♻️ refactor: 필요없는 import문 삭제 및 bookItem 컴포넌트 분리
