+++ src/views/library/components/VerticalView.tsx
@@ -1,7 +1,7 @@
 import { useState } from 'react';
 import FilterBar from './verticalview-items/FilterBar';
 import tempBookData from '../../../mock/library/bookData';
-import BookItem from './verticalview-items/BookItem';
+import BookItem from '../../../components/book-list/BookItem';
 import DeleteBtn from '../../../components/delete-button/DeleteBtn';
 
 const VerticalView = () => {
